requirements.txt
Django==4.1.1
django-richtextfield==1.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.6
python-dotenv==1.0.0
